time series graph tests moved to their own module
