+++ assets/src/scripts/components/hydrograph/index.spec.js
@@ -1,7 +1,5 @@
-import { select, selectAll } from 'd3-selection';
-import { provide } from '../../lib/redux';
-import { attachToNode, timeSeriesLegend } from './index';
-import { drawTimeSeriesGraph } from './time-series-graph';
+import { select } from 'd3-selection';
+import { attachToNode } from './index';
 import { Actions, configureStore } from '../../store';
 
 
@@ -151,7 +149,7 @@ const TEST_STATE = {
 };
 
 
-describe('Hydrograph charting module', () => {
+describe('Loading indicators and data alerts', () => {
     let graphNode;
 
     beforeEach(() => {
@@ -173,221 +171,7 @@ describe('Hydrograph charting module', () => {
         select('#hydrograph').remove();
     });
 
-    it('empty graph displays warning', () => {
-        attachToNode({}, graphNode, {});
-        expect(graphNode.innerHTML).toContain('No data is available');
-    });
-
-    it('single data point renders', () => {
-        const store = configureStore(TEST_STATE);
-        select(graphNode)
-            .call(provide(store))
-            .call(drawTimeSeriesGraph);
-        let svgNodes = graphNode.getElementsByTagName('svg');
-        expect(svgNodes.length).toBe(1);
-        expect(graphNode.innerHTML).toContain('hydrograph-container');
-    });
-
-    describe('container display', () => {
-
-        it('should not be hidden tag if there is data', () => {
-            const store = configureStore(TEST_STATE);
-            select(graphNode)
-                .call(provide(store))
-                .call(drawTimeSeriesGraph);
-            expect(select('#hydrograph').attr('hidden')).toBeNull();
-        });
-
-        it('should have a style tag if there is no data', () => {
-            const store = configureStore({series: {timeSeries: {}}});
-            select(graphNode)
-                .call(provide(store))
-                .call(drawTimeSeriesGraph);
-        });
-    });
-
-    describe('SVG has been made accessibile', () => {
-        let svg;
-        beforeEach(() => {
-            const store = configureStore(TEST_STATE);
-            select(graphNode)
-                .call(provide(store))
-                .call(drawTimeSeriesGraph);
-            svg = select('svg');
-        });
-
-        it('title and desc attributes are present', function() {
-            const descText = svg.select('desc').html();
-
-            expect(svg.select('title').html()).toEqual('Test title for 00060');
-            expect(descText).toContain('Test description for 00060');
-            expect(descText).toContain('3/23/2018');
-            expect(descText).toContain('3/30/2018');
-            expect(svg.attr('aria-labelledby')).toContain('title');
-            expect(svg.attr('aria-describedby')).toContain('desc');
-        });
-
-        it('svg should be focusable', function() {
-            expect(svg.attr('tabindex')).toBe('0');
-        });
-    });
-
-    describe('SVG contains the expected elements', () => {
-        /* eslint no-use-before-define: 0 */
-        let store;
-        beforeEach(() => {
-            store = configureStore({
-                ...TEST_STATE,
-                series: {
-                    ...TEST_STATE.series,
-                    timeSeries: {
-                        ...TEST_STATE.series.timeSeries,
-                        '00060:current': {
-                            ...TEST_STATE.series.timeSeries['00060:current'],
-                            startTime: 1514926800000,
-                            endTime: 1514930400000,
-                            points: [{
-                                dateTime: 1514926800000,
-                                value: 10,
-                                qualifiers: ['P']
-                            }, {
-                                dateTime: 1514930400000,
-                                value: null,
-                                qualifiers: ['P', 'FLD']
-                            }]
-                        }
-                    }
-                },
-                timeSeriesState: {
-                    showSeries: {
-                        current: true,
-                        compare: true,
-                        median: true
-                    },
-                    currentVariableID: '45807197',
-                    currentDateRange: 'P7D',
-                    currentMethodID: 'method1',
-                    loadingTSKeys: []
-                },
-                ui: {
-                    windowWidth: 400,
-                    width: 400
-                }
-
-            });
-
-            attachToNode(store, graphNode, {siteno: '123456788'});
-        });
-
-        it('should render the correct number of svg nodes', () => {
-            // one main hydrograph, legend and two sparklines
-            expect(selectAll('svg').size()).toBe(4);
-        });
-
-        it('should have a title div', () => {
-            const titleDiv = selectAll('.time-series-graph-title');
-            expect(titleDiv.size()).toBe(1);
-            expect(titleDiv.text()).toEqual('Test title for 00060, method description');
-        });
-
-        it('should have a defs node', () => {
-            expect(selectAll('defs').size()).toBe(1);
-            expect(selectAll('defs mask').size()).toBe(1);
-            expect(selectAll('defs pattern').size()).toBe(2);
-        });
-
-        it('should render time series data as a line', () => {
-            // There should be one segment per time-series. Each is a single
-            // point, so should be a circle.
-            expect(selectAll('.hydrograph-svg .line-segment').size()).toBe(2);
-        });
-
-        it('should render a rectangle for masked data', () => {
-            expect(selectAll('.hydrograph-svg g.current-mask-group').size()).toBe(1);
-        });
-
-        it('should have a point for the median stat data with a label', () => {
-            expect(selectAll('#median-points path').size()).toBe(1);
-            expect(selectAll('#median-points text').size()).toBe(0);
-        });
-
-        it('should have tooltips for the select series table', () => {
-            // one for each of the two parameters
-            expect(selectAll('table .tooltip-item').size()).toBe(2);
-        });
-
-        it('should not have tooltips for the select series table when the screen is large', () => {
-            store.dispatch(Actions.resizeUI(800, 800));
-            expect(selectAll('table .tooltip-table').size()).toBe(0);
-        });
-    });
-
-    //TODO: Consider adding a test which checks that the y axis is rescaled by
-    // examining the contents of the text labels.
-
-    describe('legends should render', () => {
-        let store;
-
-        beforeEach(() => {
-            store = configureStore(TEST_STATE);
-            select(graphNode)
-                .call(provide(store))
-                .call(timeSeriesLegend);
-        });
-
-        it('Should have 6 legend markers', () => {
-            expect(selectAll('.legend g').size()).toBe(6);
-            expect(selectAll('.legend g line.median-step').size()).toBe(1);
-        });
-
-        it('Should have four legend markers after the compare time series is removed', () => {
-            store.dispatch(Actions.toggleTimeSeries('compare', false));
-            expect(selectAll('.legend g').size()).toBe(4);
-        });
-
-        it('Should have two legend marker after the compare and median time series are removed', () => {
-            store.dispatch(Actions.toggleTimeSeries('compare', false));
-            store.dispatch(Actions.toggleTimeSeries('median', false));
-            expect(selectAll('.legend g').size()).toBe(2);
-        });
-    });
-
-    describe('compare line', () => {
-
-        let store;
-        beforeEach(() => {
-            store = configureStore(TEST_STATE);
-            attachToNode(store, graphNode, {siteno: '12345678'});
-        });
-
-        it('Should render one lines', () => {
-            expect(selectAll('#ts-compare-group .line-segment').size()).toBe(1);
-        });
-
-        it('Should remove the lines when removing the compare time series', () => {
-            store.dispatch(Actions.toggleTimeSeries('compare', false));
-            expect(selectAll('#ts-compare-group .line-segment').size()).toBe(0);
-        });
-    });
-
-    describe('median lines', () => {
-        let store;
-        beforeEach(() => {
-            store = configureStore(TEST_STATE);
-            attachToNode(store, graphNode, {siteno: '12345678'});
-        });
-
-        it('Should render one lines', () => {
-            expect(selectAll('#median-points .median-data-series').size()).toBe(1);
-        });
-
-        it('Should remove the lines when removing the median statistics data', () => {
-            store.dispatch(Actions.toggleTimeSeries('median', false));
-            expect(selectAll('#median-points .median-data-series').size()).toBe(0);
-        });
-    });
-
-    describe('hiding/show provisional alert', () => {
+    describe('hiding/showing provisional alert', () => {
 
         it('Expects the provisional alert to be visible when time series data is provided', () => {
             let store = configureStore(TEST_STATE);
